+++ packages/deployer/src/server/index.ts
@@ -27,7 +27,7 @@ import {
 import { handleClientsRefresh, handleTriggerClientsRefresh } from './handlers/client';
 import { errorHandler } from './handlers/error';
 import { getLogsByRunIdHandler, getLogsHandler, getLogTransports } from './handlers/logs';
-import { getMcpServerMessageHandler, handleMcpServerSseRoutes } from './handlers/mcp';
+import { getMcpServerMessageHandler, getMcpServerSseHandler } from './handlers/mcp';
 import {
   createThreadHandler,
   deleteThreadHandler,
@@ -1332,7 +1332,7 @@ export async function createHonoServer(mastra: Mastra, options: ServerBundleOpti
         500: { description: 'Internal server error establishing SSE connection.' },
       },
     }),
-    handleMcpServerSseRoutes,
+    getMcpServerSseHandler,
   );
 
   // Route for POSTing messages over an established SSE connection
@@ -1366,7 +1366,7 @@ export async function createHonoServer(mastra: Mastra, options: ServerBundleOpti
         503: { description: 'SSE connection not established with this server, or server unable to process message.' },
       },
     }),
-    handleMcpServerSseRoutes,
+    getMcpServerSseHandler,
   );
 
   // Memory routes
@@ -1468,6 +1468,13 @@ export async function createHonoServer(mastra: Mastra, options: ServerBundleOpti
           in: 'query',
           required: true,
           schema: { type: 'string' },
+        },
+        {
+          name: 'limit',
+          in: 'query',
+          required: false,
+          schema: { type: 'number' },
+          description: 'Limit the number of messages to retrieve (default: 40)',
         },
       ],
       responses: {